diplom/requirements.txt
simplejson==3.17.5
websocket_client==1.3.2
